+++ sources/org/telegram/tgnet/TLRPC$TL_channelFull_layer124.java
@@ -14,7 +14,7 @@ public class TLRPC$TL_channelFull_layer124 extends TLRPC$TL_channelFull {
         this.has_scheduled = (524288 & readInt32) != 0;
         this.can_view_stats = (1048576 & readInt32) != 0;
         this.blocked = (readInt32 & 4194304) != 0;
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = (long) abstractSerializedData.readInt32(z);
         this.about = abstractSerializedData.readString(z);
         if ((this.flags & 1) != 0) {
             this.participants_count = abstractSerializedData.readInt32(z);
@@ -53,7 +53,7 @@ public class TLRPC$TL_channelFull_layer124 extends TLRPC$TL_channelFull {
                 }
             }
             if ((this.flags & 16) != 0) {
-                this.migrated_from_chat_id = abstractSerializedData.readInt32(z);
+                this.migrated_from_chat_id = (long) abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 16) != 0) {
                 this.migrated_from_max_id = abstractSerializedData.readInt32(z);
@@ -71,7 +71,7 @@ public class TLRPC$TL_channelFull_layer124 extends TLRPC$TL_channelFull {
                 this.folder_id = abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 16384) != 0) {
-                this.linked_chat_id = abstractSerializedData.readInt32(z);
+                this.linked_chat_id = (long) abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 32768) != 0) {
                 this.location = TLRPC$ChannelLocation.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
@@ -127,7 +127,7 @@ public class TLRPC$TL_channelFull_layer124 extends TLRPC$TL_channelFull {
         int i8 = this.blocked ? i7 | 4194304 : i7 & -4194305;
         this.flags = i8;
         abstractSerializedData.writeInt32(i8);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt32((int) this.id);
         abstractSerializedData.writeString(this.about);
         if ((this.flags & 1) != 0) {
             abstractSerializedData.writeInt32(this.participants_count);
@@ -159,7 +159,7 @@ public class TLRPC$TL_channelFull_layer124 extends TLRPC$TL_channelFull {
             this.bot_info.get(i9).serializeToStream(abstractSerializedData);
         }
         if ((this.flags & 16) != 0) {
-            abstractSerializedData.writeInt32(this.migrated_from_chat_id);
+            abstractSerializedData.writeInt32((int) this.migrated_from_chat_id);
         }
         if ((this.flags & 16) != 0) {
             abstractSerializedData.writeInt32(this.migrated_from_max_id);
@@ -177,7 +177,7 @@ public class TLRPC$TL_channelFull_layer124 extends TLRPC$TL_channelFull {
             abstractSerializedData.writeInt32(this.folder_id);
         }
         if ((this.flags & 16384) != 0) {
-            abstractSerializedData.writeInt32(this.linked_chat_id);
+            abstractSerializedData.writeInt32((int) this.linked_chat_id);
         }
         if ((this.flags & 32768) != 0) {
             this.location.serializeToStream(abstractSerializedData);
