+++ sources/org/telegram/tgnet/TLRPC$TL_channelFull_layer48.java
@@ -8,7 +8,7 @@ public class TLRPC$TL_channelFull_layer48 extends TLRPC$TL_channelFull {
         this.flags = readInt32;
         int i = 0;
         this.can_view_participants = (readInt32 & 8) != 0;
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = (long) abstractSerializedData.readInt32(z);
         this.about = abstractSerializedData.readString(z);
         if ((this.flags & 1) != 0) {
             this.participants_count = abstractSerializedData.readInt32(z);
@@ -41,7 +41,7 @@ public class TLRPC$TL_channelFull_layer48 extends TLRPC$TL_channelFull {
                 }
             }
             if ((this.flags & 16) != 0) {
-                this.migrated_from_chat_id = abstractSerializedData.readInt32(z);
+                this.migrated_from_chat_id = (long) abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 16) != 0) {
                 this.migrated_from_max_id = abstractSerializedData.readInt32(z);
@@ -56,7 +56,7 @@ public class TLRPC$TL_channelFull_layer48 extends TLRPC$TL_channelFull {
         int i = this.can_view_participants ? this.flags | 8 : this.flags & -9;
         this.flags = i;
         abstractSerializedData.writeInt32(i);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt32((int) this.id);
         abstractSerializedData.writeString(this.about);
         if ((this.flags & 1) != 0) {
             abstractSerializedData.writeInt32(this.participants_count);
@@ -80,7 +80,7 @@ public class TLRPC$TL_channelFull_layer48 extends TLRPC$TL_channelFull {
             this.bot_info.get(i2).serializeToStream(abstractSerializedData);
         }
         if ((this.flags & 16) != 0) {
-            abstractSerializedData.writeInt32(this.migrated_from_chat_id);
+            abstractSerializedData.writeInt32((int) this.migrated_from_chat_id);
         }
         if ((this.flags & 16) != 0) {
             abstractSerializedData.writeInt32(this.migrated_from_max_id);
