+++ sources/org/telegram/tgnet/TLRPC$TL_channelFull_layer71.java
@@ -10,7 +10,7 @@ public class TLRPC$TL_channelFull_layer71 extends TLRPC$TL_channelFull {
         this.can_view_participants = (readInt32 & 8) != 0;
         this.can_set_username = (readInt32 & 64) != 0;
         this.can_set_stickers = (readInt32 & 128) != 0;
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = (long) abstractSerializedData.readInt32(z);
         this.about = abstractSerializedData.readString(z);
         if ((this.flags & 1) != 0) {
             this.participants_count = abstractSerializedData.readInt32(z);
@@ -46,7 +46,7 @@ public class TLRPC$TL_channelFull_layer71 extends TLRPC$TL_channelFull {
                 }
             }
             if ((this.flags & 16) != 0) {
-                this.migrated_from_chat_id = abstractSerializedData.readInt32(z);
+                this.migrated_from_chat_id = (long) abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 16) != 0) {
                 this.migrated_from_max_id = abstractSerializedData.readInt32(z);
@@ -71,7 +71,7 @@ public class TLRPC$TL_channelFull_layer71 extends TLRPC$TL_channelFull {
         int i3 = this.can_set_stickers ? i2 | 128 : i2 & -129;
         this.flags = i3;
         abstractSerializedData.writeInt32(i3);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt32((int) this.id);
         abstractSerializedData.writeString(this.about);
         if ((this.flags & 1) != 0) {
             abstractSerializedData.writeInt32(this.participants_count);
@@ -98,7 +98,7 @@ public class TLRPC$TL_channelFull_layer71 extends TLRPC$TL_channelFull {
             this.bot_info.get(i4).serializeToStream(abstractSerializedData);
         }
         if ((this.flags & 16) != 0) {
-            abstractSerializedData.writeInt32(this.migrated_from_chat_id);
+            abstractSerializedData.writeInt32((int) this.migrated_from_chat_id);
         }
         if ((this.flags & 16) != 0) {
             abstractSerializedData.writeInt32(this.migrated_from_max_id);
