+++ sources/org/telegram/tgnet/TLRPC$TL_channelFull_layer89.java
@@ -11,7 +11,7 @@ public class TLRPC$TL_channelFull_layer89 extends TLRPC$TL_channelFull {
         this.can_set_username = (readInt32 & 64) != 0;
         this.can_set_stickers = (readInt32 & 128) != 0;
         this.hidden_prehistory = (readInt32 & 1024) != 0;
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = (long) abstractSerializedData.readInt32(z);
         this.about = abstractSerializedData.readString(z);
         if ((this.flags & 1) != 0) {
             this.participants_count = abstractSerializedData.readInt32(z);
@@ -47,7 +47,7 @@ public class TLRPC$TL_channelFull_layer89 extends TLRPC$TL_channelFull {
                 }
             }
             if ((this.flags & 16) != 0) {
-                this.migrated_from_chat_id = abstractSerializedData.readInt32(z);
+                this.migrated_from_chat_id = (long) abstractSerializedData.readInt32(z);
             }
             if ((this.flags & 16) != 0) {
                 this.migrated_from_max_id = abstractSerializedData.readInt32(z);
@@ -80,7 +80,7 @@ public class TLRPC$TL_channelFull_layer89 extends TLRPC$TL_channelFull {
         int i4 = this.hidden_prehistory ? i3 | 1024 : i3 & -1025;
         this.flags = i4;
         abstractSerializedData.writeInt32(i4);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt32((int) this.id);
         abstractSerializedData.writeString(this.about);
         if ((this.flags & 1) != 0) {
             abstractSerializedData.writeInt32(this.participants_count);
@@ -107,7 +107,7 @@ public class TLRPC$TL_channelFull_layer89 extends TLRPC$TL_channelFull {
             this.bot_info.get(i5).serializeToStream(abstractSerializedData);
         }
         if ((this.flags & 16) != 0) {
-            abstractSerializedData.writeInt32(this.migrated_from_chat_id);
+            abstractSerializedData.writeInt32((int) this.migrated_from_chat_id);
         }
         if ((this.flags & 16) != 0) {
             abstractSerializedData.writeInt32(this.migrated_from_max_id);
