+++ sources/org/telegram/tgnet/TLRPC$TL_channelParticipantBanned.java
@@ -12,7 +12,7 @@ public class TLRPC$TL_channelParticipantBanned extends TLRPC$ChannelParticipant 
         }
         this.left = z2;
         this.peer = TLRPC$Peer.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
-        this.kicked_by = abstractSerializedData.readInt32(z);
+        this.kicked_by = abstractSerializedData.readInt64(z);
         this.date = abstractSerializedData.readInt32(z);
         this.banned_rights = TLRPC$TL_chatBannedRights.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
     }
@@ -23,7 +23,7 @@ public class TLRPC$TL_channelParticipantBanned extends TLRPC$ChannelParticipant 
         this.flags = i;
         abstractSerializedData.writeInt32(i);
         this.peer.serializeToStream(abstractSerializedData);
-        abstractSerializedData.writeInt32(this.kicked_by);
+        abstractSerializedData.writeInt64(this.kicked_by);
         abstractSerializedData.writeInt32(this.date);
         this.banned_rights.serializeToStream(abstractSerializedData);
     }
