+++ sources/org/telegram/tgnet/TLRPC$TL_channelParticipantBanned_layer125.java
@@ -13,8 +13,8 @@ public class TLRPC$TL_channelParticipantBanned_layer125 extends TLRPC$TL_channel
         this.left = z2;
         TLRPC$TL_peerUser tLRPC$TL_peerUser = new TLRPC$TL_peerUser();
         this.peer = tLRPC$TL_peerUser;
-        tLRPC$TL_peerUser.user_id = abstractSerializedData.readInt32(z);
-        this.kicked_by = abstractSerializedData.readInt32(z);
+        tLRPC$TL_peerUser.user_id = (long) abstractSerializedData.readInt32(z);
+        this.kicked_by = (long) abstractSerializedData.readInt32(z);
         this.date = abstractSerializedData.readInt32(z);
         this.banned_rights = TLRPC$TL_chatBannedRights.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
     }
@@ -24,8 +24,8 @@ public class TLRPC$TL_channelParticipantBanned_layer125 extends TLRPC$TL_channel
         int i = this.left ? this.flags | 1 : this.flags & -2;
         this.flags = i;
         abstractSerializedData.writeInt32(i);
-        abstractSerializedData.writeInt32(this.peer.user_id);
-        abstractSerializedData.writeInt32(this.kicked_by);
+        abstractSerializedData.writeInt32((int) this.peer.user_id);
+        abstractSerializedData.writeInt32((int) this.kicked_by);
         abstractSerializedData.writeInt32(this.date);
         this.banned_rights.serializeToStream(abstractSerializedData);
     }
