+++ sources/org/telegram/tgnet/TLRPC$TL_chat.java
@@ -16,7 +16,7 @@ public class TLRPC$TL_chat extends TLRPC$Chat {
             z2 = true;
         }
         this.call_not_empty = z2;
-        this.id = abstractSerializedData.readInt32(z);
+        this.id = abstractSerializedData.readInt64(z);
         this.title = abstractSerializedData.readString(z);
         this.photo = TLRPC$ChatPhoto.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
         this.participants_count = abstractSerializedData.readInt32(z);
@@ -48,7 +48,7 @@ public class TLRPC$TL_chat extends TLRPC$Chat {
         int i6 = this.call_not_empty ? i5 | 16777216 : i5 & -16777217;
         this.flags = i6;
         abstractSerializedData.writeInt32(i6);
-        abstractSerializedData.writeInt32(this.id);
+        abstractSerializedData.writeInt64(this.id);
         abstractSerializedData.writeString(this.title);
         this.photo.serializeToStream(abstractSerializedData);
         abstractSerializedData.writeInt32(this.participants_count);
