+++ sources/org/telegram/tgnet/TLRPC$TL_chatParticipantCreator.java
@@ -4,11 +4,11 @@ public class TLRPC$TL_chatParticipantCreator extends TLRPC$ChatParticipant {
     public static int constructor = -NUM;
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.user_id = abstractSerializedData.readInt32(z);
+        this.user_id = abstractSerializedData.readInt64(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.user_id);
+        abstractSerializedData.writeInt64(this.user_id);
     }
 }
