+++ sources/org/telegram/tgnet/TLRPC$TL_encryptedChatWaiting.java
@@ -7,8 +7,8 @@ public class TLRPC$TL_encryptedChatWaiting extends TLRPC$EncryptedChat {
         this.id = abstractSerializedData.readInt32(z);
         this.access_hash = abstractSerializedData.readInt64(z);
         this.date = abstractSerializedData.readInt32(z);
-        this.admin_id = abstractSerializedData.readInt32(z);
-        this.participant_id = abstractSerializedData.readInt32(z);
+        this.admin_id = abstractSerializedData.readInt64(z);
+        this.participant_id = abstractSerializedData.readInt64(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
@@ -16,7 +16,7 @@ public class TLRPC$TL_encryptedChatWaiting extends TLRPC$EncryptedChat {
         abstractSerializedData.writeInt32(this.id);
         abstractSerializedData.writeInt64(this.access_hash);
         abstractSerializedData.writeInt32(this.date);
-        abstractSerializedData.writeInt32(this.admin_id);
-        abstractSerializedData.writeInt32(this.participant_id);
+        abstractSerializedData.writeInt64(this.admin_id);
+        abstractSerializedData.writeInt64(this.participant_id);
     }
 }
