+++ sources/org/telegram/tgnet/TLRPC$TL_inputChannelFromMessage.java
@@ -8,13 +8,13 @@ public class TLRPC$TL_inputChannelFromMessage extends TLRPC$InputChannel {
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
         this.peer = TLRPC$InputPeer.TLdeserialize(abstractSerializedData, abstractSerializedData.readInt32(z), z);
         this.msg_id = abstractSerializedData.readInt32(z);
-        this.channel_id = abstractSerializedData.readInt32(z);
+        this.channel_id = abstractSerializedData.readInt64(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
         this.peer.serializeToStream(abstractSerializedData);
         abstractSerializedData.writeInt32(this.msg_id);
-        abstractSerializedData.writeInt32(this.channel_id);
+        abstractSerializedData.writeInt64(this.channel_id);
     }
 }
