+++ sources/org/telegram/tgnet/TLRPC$TL_inputPeerChannel.java
@@ -4,13 +4,13 @@ public class TLRPC$TL_inputPeerChannel extends TLRPC$InputPeer {
     public static int constructor = NUM;
 
     public void readParams(AbstractSerializedData abstractSerializedData, boolean z) {
-        this.channel_id = abstractSerializedData.readInt32(z);
+        this.channel_id = abstractSerializedData.readInt64(z);
         this.access_hash = abstractSerializedData.readInt64(z);
     }
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.channel_id);
+        abstractSerializedData.writeInt64(this.channel_id);
         abstractSerializedData.writeInt64(this.access_hash);
     }
 }
