+++ sources/org/telegram/tgnet/TLRPC$TL_messages_addChatUser.java
@@ -2,7 +2,7 @@ package org.telegram.tgnet;
 
 public class TLRPC$TL_messages_addChatUser extends TLObject {
     public static int constructor = -NUM;
-    public int chat_id;
+    public long chat_id;
     public int fwd_limit;
     public TLRPC$InputUser user_id;
 
@@ -12,7 +12,7 @@ public class TLRPC$TL_messages_addChatUser extends TLObject {
 
     public void serializeToStream(AbstractSerializedData abstractSerializedData) {
         abstractSerializedData.writeInt32(constructor);
-        abstractSerializedData.writeInt32(this.chat_id);
+        abstractSerializedData.writeInt64(this.chat_id);
         this.user_id.serializeToStream(abstractSerializedData);
         abstractSerializedData.writeInt32(this.fwd_limit);
     }
